Register new employee structure
Was completed the structure to register a new employee, the form has the
requires attributes that an employee must have on create.
To show the department, was necessary select all the departments and
then execute a map function to get only the department and the id, the
this object, that is an array was passed as an argument in the
options_for_select, by this way, when the HR click on the submit button,
the department was automatically linked with the employee.

In the employees_contrller was added the method to can render the 'new'
template and to create an instance for a new employee.