Create new department template
The template to create a new department is now completed, was created
the controller 'departments' to manage this actions. In this controller
was created a method to render the new template with an instance for a
new department.

Also, was created the 'new.html.rb' that containse the required for to
create a new department for @department model.

The routes was created as resoucers, due to the HR could create, update,
or delete a department.

Finally, was created the respective link to the 'new_department_path'
and was appended an icon that represents the department, in this same
template, was change the id for the 'new employee' icon from
'create-user' to 'create-employee'.