Register new employee functionality
The functionality to register a new employee was completed.

In the employees_controller was added a new method 'create', this method
has the logic to save the new employee, if the employee is saved
successfully, then a success message is showed and redirects to the root
path, otherwis an error message is showed and render the /employees/new
template.

In the model, was adedd a validation for a password, and a password
confirmation, if the password confirmation doesn't exist, an error
messages is showed and the employee is not saved in the DB.