Account functionality
To complete the account functionality, was necessary find the current
employee record in the account method due to in the same view the
employee can update her information, but, here comes some interesting.

When the employee wants to update a single record directly in the account
view, is necessary to call the update method, but the update mehtod
redirects to root path, for that reason was changed this to **request.env["HTTP_REFERER"]**
that stores the previous path every time the employee surff in the page.

Finally, in the edit.html.erb template was added an unless condition,
due to the currente employee cannot update that specific records by
himself.