Absences
It was created the absences table with the date_of_absence and
employee_id datetime fields.
This table has as FK an employee, the employee_id will take these
values.
Alongs with the model, was created an absences_controlle to with the
unique method "create", this method has the logic to register a new
absence for an employee that has not been check in yet.
The logic is as follows:
- Is getted the current date and formated to time, by this way we obtain
  the current day with the hour, minutes and seconds.
- After that, make a join query to get all the employees where the
  employee_id match with employee_id from employee_schedule table. But
  we get only those employees that doesn't check in yet, this will
  returns multiply repeated records, so we use distinct to get just once
  record per employee.
- The above query, returns an array with all requested employees, so it
  is necessary iterate thruogh all the employees. The next part, is
  register a new absence for the respective user, so, we  use employee
  as the iterative variable that has the current employee as
  employee.absences.new, where absences is the table, but it's must be
  called in plural because one employee has many absences and an absence
  belongs to an employee.
- Finally, we verify if the absence couldn't be save, then redirects to
  chekc_in_path, and shows an error message, otherwise just a success
  message is returned.

In the top of the class, an instruction is setted, that is
"protect_from_forgery with: :null_session", this is because we don't use a
form to reach the create method, and if this instruction is not setted,
an authentication error will be returned.
To call the create method, we will do automaticaly using the whenever
gem, that allows programm many tasks that will be runed in an specific
date and time.

In the employees_controller, was included the EmployeeSchedulesHelper,
that has a method that validates if an employee doesn't check in yet,
after the 9:15am, if there are many employees that doesn't check in yet,
then redirects to check in path and then the employee couldn't check in
or check out by that day.